Somehow svn merge doesn't do its job properly. Some classes are not merged, some have old code instead of new ones.

+++ edu.rice.cs.hpc.traceviewer/src/edu/rice/cs/hpc/traceviewer/painter/SpaceTimeMiniCanvas.java
@@ -175,10 +175,10 @@ public class SpaceTimeMiniCanvas extends SpaceTimeCanvas
 		Point miniBottomRight = new Point((int)topLeftPixelX+viewingWidth, (int)topLeftPixelY+viewingHeight);
 		
 		long detailTopLeftTime = (long)(miniTopLeft.x/getScaleX());
-		double detailTopLeftProcess = miniTopLeft.y/getScaleY();
+		int detailTopLeftProcess = (int) (miniTopLeft.y/getScaleY());
 		
 		long detailBottomRightTime = (long)((double)miniBottomRight.x / getScaleX());
-		double detailBottomRightProcess = miniBottomRight.y/getScaleY();
+		int detailBottomRightProcess = (int) (miniBottomRight.y/getScaleY());
 
 		stData.attributes.begProcess = detailTopLeftProcess;
 		stData.attributes.endProcess = detailBottomRightProcess;
